add rudimentary ship wayfinding

+++ core/src/dev/wizrad/solarfare/game/world/support/VectorExt.kt
@@ -1,15 +1,15 @@
 package dev.wizrad.solarfare.game.world.support
 
 import com.badlogic.gdx.math.Vector2
-import dev.wizrad.solarfare.support.cosf
-import dev.wizrad.solarfare.support.sinf
+import dev.wizrad.solarfare.support.cos
+import dev.wizrad.solarfare.support.sin
 
 class Vector {
   companion object {
     fun polar(magnitude: Float, angle: Float): Vector2 {
       return Vector2(
-        cosf(angle) * magnitude,
-        sinf(angle) * magnitude
+        cos(angle) * magnitude,
+        sin(angle) * magnitude
       )
     }
   }
